+++ src/deepiri_fuselk/helix/__init__.py
@@ -1,12 +1,18 @@
 """HELIX engine subpackage."""
 
 from deepiri_fuselk.helix.coordinate_mapper import boozer_map, field_line_pitch, q_profile
+from deepiri_fuselk.helix.disruption_filament import (
+    DisruptionFilament,
+    disruption_filament,
+    midplane_pitch_rad,
+)
 from deepiri_fuselk.helix.helical_quadtree import HQRMResult, run_hqrm
 from deepiri_fuselk.helix.helix_engine import HelixEngine, HelixResult
 from deepiri_fuselk.helix.jax_mapper import boozer_map_fast, jax_available
 from deepiri_fuselk.helix.kalman_tracker import PhaseLockedTracker, TrackerState
 
 __all__ = [
+    "DisruptionFilament",
     "HQRMResult",
     "HelixEngine",
     "HelixResult",
@@ -14,8 +20,10 @@ __all__ = [
     "TrackerState",
     "boozer_map",
     "boozer_map_fast",
+    "disruption_filament",
     "field_line_pitch",
     "jax_available",
+    "midplane_pitch_rad",
     "q_profile",
     "run_hqrm",
 ]